+++ jest.config.js
@@ -1,6 +1,5 @@
 module.exports = {
   bail: true,
-  collectCoverage: true,
   coverageThreshold: {
     global: {
       branches: 100,