+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Changed
+- Changed API. Replaced class with function [#26](https://github.com/MikeDevice/first-follow/pull/26).
+
+### Removed
+- Removed `getFirstSets`, `getFollowSets` and `getPredictSets` methods [#26](https://github.com/MikeDevice/first-follow/pull/26).
+
 ## [3.3.0] - 2020-05-02
 ### Changed
 - Changed back internal representation of `firstSets`, `followSets` and `predictSets` by Sets to arrays.
@@ -65,6 +72,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.0.0] - 2016-10-10
 
 
+[Unreleased]: https://github.com/MikeDevice/first-follow/compare/v3.3.0...HEAD
 [3.3.0]: https://github.com/MikeDevice/first-follow/compare/v3.2.1...v3.3.0
 [3.2.1]: https://github.com/MikeDevice/first-follow/compare/v3.2.0...v3.2.1
 [3.2.0]: https://github.com/MikeDevice/first-follow/compare/v3.1.0...v3.2.0